Disallow simultaneous quorum on a single key.
This is generally a problem if two callers() attempt to do
set(rev=X) and set(rev=X) and we do quorum on all the nodes simultaneously.
Sure, someone will ultimately win -- but everything will be left in a weird
state. IMO, it's better to disallow this from the start. This will also
prevent weird conditions during synchronization.