Implement zero-downtime restarts.
Now, no incoming connections will fail during a restart.
The socket is kept open and communicated to a new child when
the server is hit with a SIGHUP. We do have to drop all existing
connections however, but they will not error out just generate
a potentially spurious result.

This patch also implements safe passing of the password on the
command line.