Adding img tool for very large files.
This tool split files are some chunk boundary and uploads them
as separate blobs. It does so in a safe manner, and tracks references
to each of the blobs. In fact, this is a very basic content addressable
storage system -- automatically deduping, etc. providing that the
similarities line up at the correct boundaries.

This was built for VM images (and deltas on top of those images),
so it's highly likely to be the case here.

This also required plumbing though extra parameters for Set and
Remove which indicate whether it was successful. Good to get this
stuff out of the way now!